% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clc18}
\alias{clc18}
\title{Example Corine Land Cover (2018) polygons}
\format{
An \code{sf} object containing polygon geometries and grid identifiers:
\itemize{
\item \code{OBJECTID}: unique identificator of geometry.
\item \code{code_18}: Corine Land Cover third level class' identificator.
\item \code{Remark}: redundant field from original database.
\item \code{Area_Ha}: redundant field from original database.
\item \code{Shape_Length}: redundant field from original database.
\item \code{Shape_Area}: redundant field from original database.
\item \code{geom}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 500-m harmonization grid.
}
\usage{
clc18
}
\description{
A reduced Corine Land Cover (2018) dataset used in spatial examples.
}
\keyword{datasets}
