% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pts300_sauzeme}
\alias{pts300_sauzeme}
\title{Example 300-m grid points}
\format{
An \code{sf} object containing point geometries and grid identifiers:
\itemize{
\item \code{rinda300}: identificator of 300 m grid cell.
\item \code{X}: EPSG:3059 x-coordinate of centroid.
\item \code{Y}: EPSG:3059 y-coordinate of centroid.
\item \code{tks50}: identificator of topographic map's 50 km page.
\item \code{x}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 300-m harmonization grid.
}
\usage{
pts300_sauzeme
}
\description{
Point centres of a reduced 300-m grid used in spatial examples.
}
\keyword{datasets}
