% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pts500_sauzeme}
\alias{pts500_sauzeme}
\title{Example 500-m grid points}
\format{
An \code{sf} object containing point geometries and grid identifiers:
\itemize{
\item \code{rinda500}: identificator of 500 m grid cell.
\item \code{X}: EPSG:3059 x-coordinate of centroid.
\item \code{Y}: EPSG:3059 y-coordinate of centroid.
\item \code{tks50km}: identificator of topographic map's 50 km page.
\item \code{x}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 500-m harmonization grid.
}
\usage{
pts500_sauzeme
}
\description{
Point centres of a reduced 500-m grid used in spatial examples.
}
\keyword{datasets}
