% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tikls300_sauzeme}
\alias{tikls300_sauzeme}
\title{Example 300-m grid polygons}
\format{
An \code{sf} object containing polygon geometries and grid identifiers:
\itemize{
\item \code{rinda300}: identificator of 300 m grid cell.
\item \code{x}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 300-m harmonization grid.
}
\usage{
tikls300_sauzeme
}
\description{
A reduced 300-m grid used in spatial examples.
}
\keyword{datasets}
