% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tikls500_sauzeme}
\alias{tikls500_sauzeme}
\title{Example 500-m grid points}
\format{
An \code{sf} object containing polygon geometries and grid identifiers:
\itemize{
\item \code{rinda500}: identificator of 500 m grid cell.
\item \code{tks50km}: identificator of topographic map's 50 km page.
\item \code{x}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 500-m harmonization grid.
}
\usage{
tikls500_sauzeme
}
\description{
Point centres of a reduced 500-m grid used in spatial examples.
}
\keyword{datasets}
