% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tks93_50km}
\alias{tks93_50km}
\title{Example 50-km map tiles}
\format{
An \code{sf} object containing polygon geometries and grid identifiers:
\itemize{
\item \code{NOSAUKUMS}: name of topographic map's 50 km page.
\item \code{NUMURS}: identificator of topographic map's 50 km page.
\item \code{Shape_Length}: redundant field from original database.
\item \code{Shape_Area}: redundant field from original database.
\item \code{Shape}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 500-m harmonization grid.
}
\usage{
tks93_50km
}
\description{
A reduced 50-km grid used in spatial examples.
}
\keyword{datasets}
