% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pts1000_sauzeme}
\alias{pts1000_sauzeme}
\title{Example 1000-m grid points}
\format{
An \code{sf} object containing point geometries and grid identifiers:
\itemize{
\item \code{ID1km}: identificator of 1 km grid cell.
\item \code{tks50km}: identificator of topographic map's 50 km page.
\item \code{geometry}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 1000-m harmonization grid.
}
\usage{
pts1000_sauzeme
}
\description{
Point centres of a reduced 1000-m grid used in spatial examples.
}
\keyword{datasets}
