% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tikls1km_sauzeme}
\alias{tikls1km_sauzeme}
\title{Example 1000-m grid polygons}
\format{
An \code{sf} object containing polygon geometries and grid identifiers:
\itemize{
\item \code{yes}: indicator for location in terrestrial Latvia (all the cells contain value \code{1}).
\item \code{ID1km}: identificator of 1 km grid cell.
\item \code{tks50}: identificator of topographic map's 50 km page.
\item \code{geometry}: \verb{sf geometry} field.
}
}
\source{
Derived from a reduced section of the Latvian 1000-m harmonization grid.
}
\usage{
tikls1km_sauzeme
}
\description{
A reduced 1000-m grid used in spatial examples.
}
\keyword{datasets}
